+++ CHANGELOG.md
@@ -8,6 +8,7 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ## [0.20.0] - 2023-05-01
 
 ### Added
+* Added the `AddVerticesFrom` method for adding all vertices from another graph.
 
 ## [0.19.0] - 2023-04-23